+++ database/factories/UserFactory.php
@@ -1,0 +1,21 @@
+<?php
+
+namespace Database\Factories;
+
+use App\Models\User;
+use Illuminate\Database\Eloquent\Factories\Factory;
+use Illuminate\Support\Facades\Hash;
+
+class UserFactory extends Factory
+{
+    protected $model = User::class;
+
+    public function definition(): array
+    {
+        return [
+            'name'     => $this->faker->name(),
+            'username' => $this->faker->unique()->userName(),
+            'password' => Hash::make('password'),
+        ];
+    }
+}