+++ database/seeders/DatabaseSeeder.php
@@ -9,6 +9,7 @@ class DatabaseSeeder extends Seeder
     public function run(): void
     {
         $this->call([
+            UserSeeder::class,
             DriverSeeder::class,
             BusSeeder::class,
             TripSeeder::class,